add router, repository and presentacion

+++ app/src/main/java/com/example/a2doparcial/presentacion/clima/ClimaEstado.kt
@@ -3,9 +3,9 @@ package com.example.a2doparcial.presentacion.clima
 sealed class ClimaEstado {
     data class Exitoso (
         val ciudad: String = "",
-        val temperatura: Int = 0,
+        val temperatura: Double = 0.00,
         val descripcion: String= "",
-        val st :Int = 0,
+        val st :Double = 0.00,
     ) : ClimaEstado()
     data class Error(
         val mensaje :String = "",
